// src/app/models/questionnaire.model.ts
export class Questionnaire {
  questionId: number;
  subQuestion: number;
  correctValue: string;
  status: string;
  ANS1: string;
  v_isalch: string;
  response: string;
  errorMessage: string;
  pregnancyStatus: string;
  alcoholismStatus: string;
  ans_de1: string;
  correct: number;
  memberName: string;
  questionDescription: string;
  formQuestionNumber: string;
  memberNumber: string;
  answerDescription: string;
  answerType: string;
  contract_id: string;
  partition_no: string;
  version: string;
  part_id: string;
  answer: string;
  description: string;
  partner_no: string;
  question_no: string;
  appln_no: string;
  member_no: string;
  de_flag: string;
  questionText: string;
  criteria: string;
  options: string[];
  dependentQuestions: Questionnaire[];
  applicationNumber: string;
  insuredPersonNumber: string;
  contractID: string;
  mandatoryStatus: string;
  passportProof: string;
  nationality: string;
  ipNationality: string;
  propertyType: string;

  constructor() {
    this.questionId = 0;
    this.subQuestion = 0;
    this.correctValue = '';
    this.status = '';
    this.ANS1 = '';
    this.v_isalch = '';
    this.response = '';
    this.errorMessage = '';
    this.pregnancyStatus = '';
    this.alcoholismStatus = '';
    this.ans_de1 = '';
    this.correct = 0;
    this.memberName = '';
    this.questionDescription = '';
    this.formQuestionNumber = '';
    this.memberNumber = '';
    this.answerDescription = '';
    this.answerType = '';
    this.contract_id = '';
    this.partition_no = '';
    this.version = '';
    this.part_id = '';
    this.answer = '';
    this.description = '';
    this.partner_no = '';
    this.question_no = '';
    this.appln_no = '';
    this.member_no = '';
    this.de_flag = '';
    this.questionText = '';
    this.criteria = '';
    this.options = [];
    this.dependentQuestions = [];
    this.applicationNumber = '';
    this.insuredPersonNumber = '';
    this.contractID = '';
    this.mandatoryStatus = '';
    this.passportProof = '';
    this.nationality = '';
    this.ipNationality = '';
    this.propertyType = '';
  }
}